Introduce custom session authentications
Provide the interface for session authenticator. Allows clients
to perform custom authentication step right before session is
attached to the router. This allows to perform per-router custom
checks of the session to be established.
This is not a part of WAMP specification. To include this
functionality one needs to specify CMake variable
BONEFISH_USE_SESSION_AUTHENTICATOR along with compile time
definition with the same name for the code using this functionality.